pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mindbuddy"
version = "0.1.0"
description = "A privacy-focused journaling buddy that helps users reflect on stress, habits, and moods"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "MindBuddy Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business :: News/Diary",
]
dependencies = [
    "llama-index>=0.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.23.0",
    "slowapi>=0.1.8",
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
]

[project.scripts]
mindbuddy-cli = "src.mindbuddy.cli:main"
mindbuddy-api = "src.mindbuddy.api:run_server"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py313"
line-length = 100
include = ["**/*.py"] 

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "D", "N"]
ignore = ["D203", "D213"]  # Removes the conflicting docstring rules
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["waypoint_rag_eval"]
combine-as-imports = true
